Changed endpoint location to table file
