+++ backend/swagger.js
@@ -1,7 +1,7 @@
 const swaggerJsdoc = require('swagger-jsdoc');
 
 const options = {
-  apis: ['./routes/restaurant.js', './routes/user.js', './routes/reservation.js'],
+  apis: ['./routes/restaurant.js', './routes/user.js', './routes/reservation.js', './routes/table.js'],
   basePath: '/',
   swaggerDefinition: {
     info: {
